+++ project/components/columns/TeamMemberColumns.tsx
@@ -13,11 +13,11 @@ import {
     DropdownMenuTrigger,
 } from "@/components/ui/dropdown-menu";
 import { Button } from "@/components/ui/button";
-import KickMember from "@/components/dialog/KickMember";
+import KickMember from "@/components/modals/KickMember";
 import { useState } from "react";
 import { useAuthRoleByTeam } from "@/hooks/useRoles";
 import { Avatar, AvatarFallback, AvatarImage } from "@/components/ui/avatar";
-import UpdateMember from "../dialog/UpdateMember";
+import UpdateMember from "../modals/UpdateMember";
 
 export const TeamMemberColumns: (
     teamSlug: string
